# public/openapi.yaml
openapi: 3.1.0
info:
  title: Kaspi Automation API
  version: 1.0.0
  description: Tenant-isolated API for Kaspi QR payments, phone invoices, recurring schedules, cancellation, refunds, and safe sandbox simulation.
servers:
  - url: /api/v1
security:
  - bearerApiKey: []
paths:
...
  /payments:
    post:
      operationId: createPayment
      summary: Create a QR payment or phone invoice
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePayment' }
      responses:
        '201': { $ref: '#/components/responses/Payment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/ProviderError' }
    get:
      operationId: listPayments
      summary: List recent payments
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Payment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /payments/{paymentId}:
    get:
      operationId: getPayment
      summary: Get a payment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200': { $ref: '#/components/responses/Payment' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePaymentNote
      summary: Update the private internal note
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [internalComment]
              properties:
                internalComment: { type: [string, 'null'], maxLength: 255 }
      responses:
        '200': { $ref: '#/components/responses/Payment' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: kpa_live
      description: API key created in the customer dashboard.
  parameters:
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key retained for 24 hours. Reusing it with different input returns 409.
      schema: { type: string, minLength: 8, maxLength: 200 }
...
  schemas:
...
    CreatePayment:
      type: object
      required: [method, amount]
      properties:
        method: { type: string, enum: [qr, invoice] }
        amount: { type: number, exclusiveMinimum: 0, maximum: 999999999 }
        currency: { type: string, const: KZT, default: KZT }
        externalOrderId: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 500 }
        customerPhone:
          type: string
          pattern: '^7?\d{10}$'
          description: Required when method is invoice.
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        metadata: { type: object, additionalProperties: true }
    Payment:
      type: object
      required: [id, method, amount, currency, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        method: { type: string, enum: [qr, invoice] }
        amount: { type: number }
        currency: { type: string, const: KZT }
        externalOrderId: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        customerPhone: { type: [string, 'null'] }
        status:
          type: string
          enum: [created, requires_customer_action, pending, paid, failed, expired, cancelled, unknown, partially_refunded, refunded]
        providerStatus: { type: [string, 'null'] }
        qrToken: { type: [string, 'null'], format: uri }
        qrOriginalToken: { type: [string, 'null'], format: uri }
        qrCodeDataUrl: { type: [string, 'null'], description: Present on QR payment creation responses. }
        receiptUrl: { type: [string, 'null'], format: uri }
        metadata: { type: object, additionalProperties: true }
        expiresAt: { type: [string, 'null'], format: date-time }
        paidAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        requestId: { type: string }
  responses:
    Payment:
      description: Payment
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: '#/components/schemas/Payment' }
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: API key missing or invalid
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: API key scope is insufficient
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found in this tenant
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: State or idempotency conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ProviderError:
      description: Kaspi rejected or could not complete the operation
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...